% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_raster}
\alias{example_raster}
\title{Example raster}
\usage{
example_raster()
}
\value{
A single-layer \link[terra:SpatRaster]{terra::SpatRaster} object.
}
\description{
Returns a small single-layer \link[terra:SpatRaster]{terra::SpatRaster} with 16 rows and 20 columns
of known values covering the extent of \code{\link[=example_polygons]{example_polygons()}}.
Includes some \code{NA} cells in the upper-right corner and valid zero values.
}
\details{
The raster uses EPSG:32632 (UTM zone 32N) with bounding box
\verb{xmin = 398000, xmax = 432000, ymin = 5398000, ymax = 5422000}
(1700 m in X by 1500 m in Y cell resolution).
}
\examples{
example_raster()
}
